Got build to work with server and client we did it :D I'm so tired need sleep
