+++ ORIGINAL_README.md
@@ -18,9 +18,9 @@ You can reach the test server at [http://localhost:3000](http://localhost:3000)
 
 ## Code structure
 
-All your code must be in the `public` folder. Put your server side code into the `server.js` file. The `shared.js` file is loaded at the begining of the `server.js` file. You can also use this code on the client side.
+All your code must be in the `public` folder. Put your server side code into the `server.old.js` file. The `shared.js` file is loaded at the begining of the `server.old.js` file. You can also use this code on the client side.
 
-The `server.js` is a standard Node.js module. You can use the following structure to create new [Express](https://expressjs.com/) routes or [Socket.io](https://socket.io/) connection handler.
+The `server.old.js` is a standard Node.js module. You can use the following structure to create new [Express](https://expressjs.com/) routes or [Socket.io](https://socket.io/) connection handler.
 
     module.exports = {
         // Express route to /hello
@@ -35,7 +35,7 @@ The server category entries can use 13kByte persistent storage. The key and valu
 
 The storage uses SQL database to save the key/value pairs. By default it's SQLite but on Heroku you have to use the Postgres add-on!
 
-In the `server.js` file you can access the `storage` interface. The API documentation is inside the [lib](./lib/) folder.
+In the `server.old.js` file you can access the `storage` interface. The API documentation is inside the [lib](./lib/) folder.
 
 ## Deploy to Heroku
