+++ package.json
@@ -3,11 +3,11 @@
   "version": "1.0.2",
   "description": "js13kgames game jam server",
   "license": "MIT",
-  "main": "index.js",
+  "main": "server/index.js",
   "scripts": {
-    "start": "node index.js",
+    "start": "node ./public/server.js",
     "test": "mocha test",
-    "build": "gulp"
+    "build": "(cd ./server; webpack)"
   },
   "keywords": [
     "js13kgames",
@@ -32,15 +32,19 @@
     "gulp-uglify": "^3.0.1",
     "gulp-zip": "^4.2.0",
     "mysql2": "^1.5.3",
+    "napa": "3.0.0",
     "pg": "^7.4.3",
     "pg-hstore": "^2.3.2",
     "sequelize": "^4.38.0",
-    "socket.io": "^2.1.0",
-    "sqlite3": "^4.0.1",
+    "socket.io": "^2.2.0",
+    "socket.io-client": "^2.2.0",
+    "sqlite3": "*",
     "uglify-js": "^3.4.9",
     "underscore": "^1.9.1",
+    "uws": "10.148.1",
     "webpack": "^4.25.1",
-    "webpack-cli": "^3.1.2"
+    "webpack-cli": "^3.1.2",
+    "webpack-node-externals": "^1.7.2"
   },
   "devDependencies": {
     "mocha": "^5.2.0"
