+++ server/Level/Level.js
@@ -1,0 +1,10 @@
+//TODO: Do something with this jazz
+
+const map_1 = "[{\"type\":\"Starting\",\"args\":[50,40,50,800,1050,800,1050,40]},{\"type\":\"Floor\",\"args\":[550,420,790,1040]},{\"type\":\"Wall\",\"args\":[30,420,820,20]},{\"type\":\"Wall\",\"args\":[550,820,20,1035]},{\"type\":\"Wall\",\"args\":[1075,420,820,20]},{\"type\":\"Wall\",\"args\":[552,15,20,1065]},{\"type\":\"Wall\",\"args\":[150,400,20,250]},{\"type\":\"Wall\",\"args\":[945,400,20,240]},{\"type\":\"Wall\",\"args\":[540,680,260,20]},{\"type\":\"Wall\",\"args\":[540,150,260,20]},{\"type\":\"Wall\",\"args\":[120,165,130,20]},{\"type\":\"Wall\",\"args\":[975,165,130,20]},{\"type\":\"Wall\",\"args\":[920,100,20,130]},{\"type\":\"Wall\",\"args\":[175,100,20,130]},{\"type\":\"Wall\",\"args\":[120,680,130,20]},{\"type\":\"Wall\",\"args\":[975,680,130,20]},{\"type\":\"Wall\",\"args\":[180,735,20,140]},{\"type\":\"Wall\",\"args\":[915,735,20,140]},{\"type\":\"GroundShotgun\",\"args\":[150,136]},{\"type\":\"GroundShotgun\",\"args\":[900,130]},{\"type\":\"GroundShotgun\",\"args\":[900,670]},{\"type\":\"GroundShotgun\",\"args\":[150,670]},{\"type\":\"GroundMachineGun\",\"args\":[536,400]}]";
+const map_2 = "[{\"type\":\"Starting\",\"args\":[50,40,50,800,1050,800,1050,40]},{\"type\":\"Wall\",\"args\":[30,420,820,20]},{\"type\":\"Wall\",\"args\":[550,820,20,1035]},{\"type\":\"Wall\",\"args\":[1075,420,820,20]},{\"type\":\"Wall\",\"args\":[552,15,20,1065]},{\"type\":\"Wall\",\"args\":[410,250,280,20]},{\"type\":\"Wall\",\"args\":[670,250,280,20]},{\"type\":\"Wall\",\"args\":[540,555,20,575]},{\"type\":\"Wall\",\"args\":[265,477,137,20]},{\"type\":\"Wall\",\"args\":[815,477,137,20]},{\"type\":\"GroundSmg\",\"args\":[480,420]},{\"type\":\"GroundSmg\",\"args\":[570,420]},{\"type\":\"GroundShotgun\",\"args\":[290,500]},{\"type\":\"GroundShotgun\",\"args\":[750,500]}]";
+const map_3 = "[{\"type\":\"Starting\",\"args\":[550,240,400,380,700,380,550,530]},{\"type\":\"Wall\",\"args\":[30,420,820,20]},{\"type\":\"Wall\",\"args\":[550,820,20,1035]},{\"type\":\"Wall\",\"args\":[1075,420,820,20]},{\"type\":\"Wall\",\"args\":[552,15,20,1065]},{\"type\":\"Wall\",\"args\":[550,385,260,280]},{\"type\":\"GroundShotgun\",\"args\":[55,35]},{\"type\":\"GroundShotgun\",\"args\":[55,750]},{\"type\":\"GroundShotgun\",\"args\":[1000,750]},{\"type\":\"GroundShotgun\",\"args\":[1000,35]}]";
+const map_4 = "[{\"type\":\"Starting\",\"args\":[110,50,260,50,850,50,980,50]},{\"type\":\"Wall\",\"args\":[30,420,820,20]},{\"type\":\"Wall\",\"args\":[550,820,20,1035]},{\"type\":\"Wall\",\"args\":[1075,420,820,20]},{\"type\":\"Wall\",\"args\":[552,15,20,1065]},{\"type\":\"Wall\",\"args\":[200,372,700,20]},{\"type\":\"Wall\",\"args\":[565,460,705,20]},{\"type\":\"Wall\",\"args\":[905,372,700,20]},{\"type\":\"GroundMachineGun\",\"args\":[885,745]},{\"type\":\"GroundMachineGun\",\"args\":[180,745]},{\"type\":\"GroundSmg\",\"args\":[540,70]}]";
+const map_5 = "[{\"type\":\"Starting\",\"args\":[145,370,415,370,712,370,960,370]},{\"type\":\"Wall\",\"args\":[30,420,820,20]},{\"type\":\"Wall\",\"args\":[550,820,20,1035]},{\"type\":\"Wall\",\"args\":[1075,420,820,20]},{\"type\":\"Wall\",\"args\":[552,15,20,1065]},{\"type\":\"Wall\",\"args\":[270,372,700,20]},{\"type\":\"Wall\",\"args\":[565,460,705,20]},{\"type\":\"Wall\",\"args\":[840,372,700,20]},{\"type\":\"GroundMachineGun\",\"args\":[820,745]},{\"type\":\"GroundMachineGun\",\"args\":[245,745]},{\"type\":\"GroundMachineGun\",\"args\":[550,40]}]";
+const map_6 = "[{\"type\":\"Starting\",\"args\":[520,365,580,365,520,450,580,450]},{\"type\":\"Wall\",\"args\":[30,420,820,20]},{\"type\":\"Wall\",\"args\":[550,820,20,1035]},{\"type\":\"Wall\",\"args\":[1075,420,820,20]},{\"type\":\"Wall\",\"args\":[552,15,20,1065]},{\"type\":\"Wall\",\"args\":[555,200,20,480]},{\"type\":\"Wall\",\"args\":[305,396,412,20]},{\"type\":\"Wall\",\"args\":[805,396,412,20]},{\"type\":\"Wall\",\"args\":[555,612,20,520]},{\"type\":\"GroundShotgun\",\"args\":[340,200]},{\"type\":\"GroundShotgun\",\"args\":[720,200]},{\"type\":\"GroundShotgun\",\"args\":[340,550]},{\"type\":\"GroundShotgun\",\"args\":[720,550]}]";
+const map_7 = "[{\"type\":\"Starting\",\"args\":[280,180,800,280,280,740,850,740]},{\"type\":\"Wall\",\"args\":[30,420,820,20]},{\"type\":\"Wall\",\"args\":[550,820,20,1035]},{\"type\":\"Wall\",\"args\":[1075,420,820,20]},{\"type\":\"Wall\",\"args\":[552,15,20,1065]},{\"type\":\"Wall\",\"args\":[535,393,500,20]},{\"type\":\"Wall\",\"args\":[540,370,20,780]},{\"type\":\"GroundSmg\",\"args\":[150,136]},{\"type\":\"GroundShotgun\",\"args\":[900,130]},{\"type\":\"GroundSmg\",\"args\":[900,670]},{\"type\":\"GroundShotgun\",\"args\":[150,670]},{\"type\":\"GroundMachineGun\",\"args\":[560,400]}]";
+const map_8 = "[{\"type\":\"Starting\",\"args\":[50,40,50,800,1050,800,1050,40]},{\"type\":\"Floor\",\"args\":[550,420,790,1040]},{\"type\":\"Wall\",\"args\":[30,420,820,20]},{\"type\":\"Wall\",\"args\":[550,820,20,1035]},{\"type\":\"Wall\",\"args\":[1075,420,820,20]},{\"type\":\"Wall\",\"args\":[552,15,20,1065]},{\"type\":\"Wall\",\"args\":[150,400,20,250]},{\"type\":\"Wall\",\"args\":[945,400,20,240]},{\"type\":\"Wall\",\"args\":[540,680,260,20]},{\"type\":\"Wall\",\"args\":[540,150,260,20]},{\"type\":\"Wall\",\"args\":[120,165,130,20]},{\"type\":\"Wall\",\"args\":[975,165,130,20]},{\"type\":\"Wall\",\"args\":[920,100,20,130]},{\"type\":\"Wall\",\"args\":[175,100,20,130]},{\"type\":\"Wall\",\"args\":[120,680,130,20]},{\"type\":\"Wall\",\"args\":[975,680,130,20]},{\"type\":\"Wall\",\"args\":[180,735,20,140]},{\"type\":\"Wall\",\"args\":[915,735,20,140]},{\"type\":\"GroundSmg\",\"args\":[150,136]},{\"type\":\"GroundSmg\",\"args\":[900,130]},{\"type\":\"GroundSmg\",\"args\":[900,670]},{\"type\":\"GroundSmg\",\"args\":[150,670]},{\"type\":\"GroundMachineGun\",\"args\":[536,400]}]";
