+++ server/index.js
@@ -1,0 +1,10 @@
+import Cthulu from "./Cthulu/Cthulu";
+
+let cthulu = new Cthulu();
+
+cthulu.awaken();
+
+
+
+
+
